# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestor Recursos API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:80/api/v1"
        }
    ],
    "paths": {
...
        "/webpage?searchTags=[]&searchNombre=''&page=1&perPage=5": {
            "get": {
                "tags": [
                    "WebPages"
                ],
                "summary": "List WebPages",
                "description": "List WebPages",
                "operationId": "getWebPages",
                "parameters": [
                    {
                        "name": "searchTags",
                        "in": "query",
                        "description": "Array of Tag's id",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "number"
                            }
                        }
                    },
                    {
                        "name": "searchNombre",
                        "in": "query",
                        "description": "Name to filter recourses",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Number of page of pagination",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Amount of register for page",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "identificador": {
                                                        "type": "integer",
                                                        "example": 8
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Developer Roadmaps"
                                                    },
                                                    "url": {
                                                        "type": "string",
                                                        "example": "https://roadmap.sh/"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "RoadMaps para formación profesional de Desarrolladores"
                                                    },
                                                    "totalVisitas": {
                                                        "type": "integer",
                                                        "example": 0
                                                    },
                                                    "tags": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "identificador": {
                                                                    "type": "integer",
                                                                    "example": 2
                                                                },
                                                                "nombre": {
                                                                    "type": "string",
                                                                    "example": "PROGRAMACIÓN"
                                                                },
                                                                "estilos": {
                                                                    "type": "string",
                                                                    "example": "bg-gray-900 text-white"
                                                                },
                                                                "total": {
                                                                    "type": "integer",
                                                                    "example": 0
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag"
                                                },
                                                "currentPage": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "perPage": {
                                                    "type": "number",
                                                    "example": 5
                                                },
                                                "totalPages": {
                                                    "type": "number",
                                                    "example": 3
                                                },
                                                "from": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "to": {
                                                    "type": "number",
                                                    "example": 5
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "example": 15
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "links": {
                                            "properties": {
                                                "self": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=1"
                                                },
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=4"
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://localhost/api/v1/tag?perPage=5&page=2"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/webpage": {
            "post": {
                "tags": [
                    "WebPages"
                ],
                "summary": "Save a WebPages",
                "description": "Save a WebPages",
                "operationId": "postWebPage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "url",
                                    "description",
                                    "tags"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mi Etiqueta"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "Mi Etiqueta"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Mi Etiqueta"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        },
                                        "example": "[]"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "miweb",
                                    "url": "http://web.com",
                                    "description": "Web Testing",
                                    "tags": "[]"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": 201
                                        },
                                        "data": {
                                            "properties": {
                                                "identificador": {
                                                    "type": "integer",
                                                    "example": 8
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Developer Roadmaps"
                                                },
                                                "url": {
                                                    "type": "string",
                                                    "example": "https://roadmap.sh/"
                                                },
                                                "descripcion": {
                                                    "type": "string",
                                                    "example": "RoadMaps para formación profesional de Desarrolladores"
                                                },
                                                "totalVisitas": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "tags": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "identificador": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "PROGRAMACIÓN"
                                                            },
                                                            "estilos": {
                                                                "type": "string",
                                                                "example": "bg-gray-900 text-white"
                                                            },
                                                            "total": {
                                                                "type": "integer",
                                                                "example": 0
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}